+++ src/main/java/org/mattlang/jc/engine/search/MoveScore.java
@@ -7,9 +7,12 @@ public class MoveScore {
     public final int move;
     public final int score;
 
-    public MoveScore(int move, int score) {
+    public final int numSearchedNodes;
+
+    public MoveScore(int move, int score, int numSearchedNodes) {
         this.move = move;
         this.score = score;
+        this.numSearchedNodes=numSearchedNodes;
     }
 
     @Override