+++ src/main/java/org/mattlang/jc/engine/sorting/LongSorter.java
@@ -68,10 +68,10 @@ public class LongSorter {
         int currLowest = -1;
         int currLowestIndex = -1;
         for (int i = start; i < size - 1; i++) {
-            if (orders[i] > orders[i + 1]) {
-                swap(i, i + 1);
-
-            }
+//            if (orders[i] > orders[i + 1]) {
+//                swap(i, i + 1);
+//
+//            }
             if (orders[i] < currLowest || currLowest == -1) {
                 currLowest = orders[i];
                 currLowestIndex = i;
@@ -82,9 +82,9 @@ public class LongSorter {
         if (currLowestIndex != start) {
             swap(start, currLowestIndex);
         }
-        if (swapCounter == 0) {
-            alreadyFullySorted = true;
-        }
+//        if (swapCounter == 0) {
+//            alreadyFullySorted = true;
+//        }
     }
 
     private void swap(int i, int j) {