+++ src/main/java/org/mattlang/jc/engine/sorting/OrderCalculator.java
@@ -5,11 +5,11 @@ import java.util.HashMap;
 import org.mattlang.jc.Factory;
 import org.mattlang.jc.board.BoardRepresentation;
 import org.mattlang.jc.board.Color;
+import org.mattlang.jc.board.FigureType;
 import org.mattlang.jc.board.Move;
 import org.mattlang.jc.engine.search.CounterMoveHeuristic;
 import org.mattlang.jc.engine.search.HistoryHeuristic;
 import org.mattlang.jc.engine.search.KillerMoves;
-import org.mattlang.jc.engine.search.MoveScore;
 import org.mattlang.jc.engine.see.SEE;
 
 import lombok.Getter;
@@ -22,6 +22,8 @@ public class OrderCalculator {
     public static final int GOOD_CAPTURES_SCORE = -100_000_000;
 
     public static final int KILLER_SCORE = -10_000_000;
+
+    public static final int QUEEN_PROMOTION_SCORE = -80_000_000;
 
     public static final int COUNTER_MOVE_SCORE = -5_000_000;
 
@@ -73,16 +75,16 @@ public class OrderCalculator {
         this.parentMove = parentMove;
         int index = ply - 1;
         // if we are at the root and have scores from a previous run, lets take them:
-        if (index == 0 && orderHints.moveScores != null) {
-            // todo assert that the first pvs should be the highest score...
-            scores = new HashMap<>();
-            for (MoveScore moveScore : orderHints.moveScores) {
-                scores.put(moveScore.move, moveScore.score);
-            }
-
-        } else {
+//        if (index == 0 && orderHints.moveScores != null) {
+//            // todo assert that the first pvs should be the highest score...
+//            scores = new HashMap<>();
+//            for (MoveScore moveScore : orderHints.moveScores) {
+//                scores.put(moveScore.move, moveScore.numSearchedNodes);
+//            }
+//
+//        } else {
             scores = null;
-        }
+//        }
 
         this.ply = ply;
         this.pvMove = orderHints.prevPvlist != null ? orderHints.prevPvlist.getMove(index) : 0;
@@ -115,20 +117,19 @@ public class OrderCalculator {
      */
     public int calcOrder(Move m) {
         int moveInt = m.toInt();
-        if (scores != null) {
-            Integer rslt = scores.get(moveInt);
-            if (rslt == null){
-               // UCILogger.log("hey!! this should not happen!!!! cant find scores for " + m.toStr());
-                // this could happen for pseudo legal moves for illegal moves: here we have no score from the last round
-                return 0;
-            }
-            return -rslt;
-        }
-
         if (usePvSorting && pvMove == moveInt) {
             return PV_SCORE;
         } else if (hashMove == moveInt) {
             return HASHMOVE_SCORE;
+        } else if (scores != null) {
+            Integer rslt = scores.get(moveInt);
+            if (rslt == null) {
+                // UCILogger.log("hey!! this should not happen!!!! cant find scores for " + m.toStr());
+                // this could happen for pseudo legal moves for illegal moves: here we have no score from the last round
+                return 0;
+            }
+            return -rslt;
+
         } else {
             int mvvLva = useMvvLva ? MvvLva.calcMMVLVA(m) : 0;
 
@@ -142,6 +143,8 @@ public class OrderCalculator {
 
             } else if (killerMoves != null && killerMoves.isKiller(color, moveInt, ply)) {
                 return KILLER_SCORE;
+            } else if (m.isPromotion() && m.getPromotedFigure().figureType == FigureType.Queen) {
+               return QUEEN_PROMOTION_SCORE;
             } else if (counterMoveHeuristic != null
                     && counterMoveHeuristic.getCounter(color.ordinal(), parentMove) == moveInt) {
                 return COUNTER_MOVE_SCORE;